backend/requirements.txt
fastapi[standard]
uvicorn[standard]
python-dotenv
SQLModel
pytest
# auth / security
 passlib[bcrypt]
 python-jose[cryptography]